requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
psycopg2-binary==2.9.11
python-dotenv==1.0.0
requests==2.31.0
qrcode[pil]==7.4.2
